// apps/meet/src/store/slices/interfaces/room-settings.ts
import { VideoQuality } from 'livekit-client';
import { Theme } from '@excalidraw/excalidraw/element/types';
import type { TypeOptions } from 'react-toastify';

export enum VideoObjectFit {
  COVER = 'cover',
  CONTAIN = 'contain',
}

export interface IRoomSettings {
  isShowRoomSettingsModal: boolean;
  isNatsServerConnected: boolean;

  audioDevices: Array<IMediaDevice>;
  videoDevices: Array<IMediaDevice>;
  selectedAudioDevice: string;
  selectedVideoDevice: string;
  playAudioNotification: boolean;
  activateWebcamsView: boolean;
  activeScreenSharingView: boolean;
  allowPlayAudioNotification: boolean;
  roomAudioVolume: number;
  roomScreenShareAudioVolume: number;
  roomVideoQuality: VideoQuality;
  theme: Theme;
  videoObjectFit: VideoObjectFit;

  selectedTabLeftPanel: number;
  selectedChatOption: string;
  initiatePrivateChat: InitiatePrivateChat;
  unreadMsgFrom: Array<string>;
  selectedChatTransLang: string;

  columnCameraWidth: ColumnCameraWidth;
  columnCameraPosition: ColumnCameraPosition;
  visibleHeader: boolean;
  visibleFooter: boolean;
  isWAJLCWindowTabVisible: boolean;
  pinCamUserId?: string;
  focusActiveSpeakerWebcam: boolean;
  userNotifications: UserNotification[];
  isSidePanelOpened: boolean;
  hasWebcamPages: boolean;
  maxNumDisplayWebcams?: number;
}

export interface IMediaDevice {
  id: string;
  label: string;
}

export interface InitiatePrivateChat {
  name: string;
  userId: string;
}

export interface UnreadMsgFromPayload {
  task: 'ADD' | 'DEL';
  id: string;
}

export enum ColumnCameraWidth {
  FULL_WIDTH = 'full',
  MEDIUM_WIDTH = 'medium',
  SMALL_WIDTH = 'small',
}

export enum ColumnCameraPosition {
  LEFT = 'left',
  TOP = 'top',
  BOTTOM = 'bottom',
}

export interface UserNotification {
  message: string;
  typeOption: TypeOptions;
  notificationCat?: NotificationCats;
  data?: string;
  newInstance?: boolean;
  autoClose?: number | false;
  created?: number;
  disableToastNotification?: boolean;
}

export type NotificationCats =
    | 'new-poll-created'
    | 'breakout-room-invitation'
    | 'default';